trl_ppo/requirements.txt
torch>=2.0.0
transformers>=4.40.0
datasets>=2.14.0
peft>=0.10.0
trl>=0.8.6
huggingface_hub>=0.17.0
pyyaml>=6.0
tqdm>=4.66.0
sentencepiece>=0.1.99
protobuf>=3.20.0
accelerate>=0.28.0
